+++ src/main/java/life/majiang/community/dto/ManagerDTO.java
@@ -14,6 +14,6 @@ public class ManagerDTO {
     private Integer pay;
     private String phone;
     private Integer department;
-    private BigInteger person;
+    private Integer person;
     private User user;
 }
